Moved determineAffectedPackges to it's own repo
