+++ utils/index.ts
@@ -2,4 +2,5 @@ export { ansiColors as colors } from "./src/ansiColors";
 export { listAllPkgJsons } from "./src/listAllPkgJsons";
 export { logIterable } from "./src/logIterable";
 export { install } from "./src/install";
-
+export type { LernaListOutputType } from "./src/detectAffectedPackages";
+export { detectAffectedPackages } from "./src/detectAffectedPackages";
